+++ backend/cmd/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	_ "github.com/lib/pq"
 	"github.com/yoshihiro-shu/draft-backend/config"
-	"github.com/yoshihiro-shu/draft-backend/server"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server"
 )
 
 func main() {
